backend/sde1prep/app/requirements.txt
fastapi
uvicorn
motor
pydantic-settings
python-jose[cryptography]
python-multipart
email-validator
bcrypt==3.2.2